requirements.txt
qrcode~=7.4.2
pyserial~=3.5
PyQt5~=5.15.9
Pillow~=10.0.0